services/api/requirements.txt
fastapi==0.115.6
asyncpg==0.30.0
httpx==0.27.2
alembic==1.14.1
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
greenlet==3.0.3
prometheus_client==0.21.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-httpx==0.51b0
uvicorn[standard]==0.34.0
google-auth==2.27.0
requests==2.32.3
pyjwt[crypto]==2.10.1
